Ensure file exists before attempting to move to xdg

+++ src/main/java/com/atlauncher/FileSystem.java
@@ -178,8 +178,10 @@ public final class FileSystem {
 
     private static void filesMove(Path a, Path b) throws IOException {
         System.out.println("Moving " + a + " -> " + b);
-        FileUtils.moveFile(a, b);
-        FileUtils.delete(a);
+        if (a.toFile().exists()) {
+            FileUtils.moveFile(a, b);
+            FileUtils.delete(a);
+        }
     }
 
     /**
